+++ web/src/typeahead_helper.ts
@@ -807,6 +807,10 @@ function compare_by_name(stream_a: StreamSubscription, stream_b: StreamSubscript
     return util.strcmp(stream_a.name, stream_b.name);
 }
 
+function compare_by_user_group_name(group_a: UserGroup, group_b: UserGroup): number {
+    return util.strcmp(group_a.name, group_b.name);
+}
+
 export let sort_streams = (matches: StreamPillData[], query: string): StreamPillData[] => {
     const name_results = typeahead.triage(query, matches, (x) => x.name, compare_by_activity);
     const desc_results = typeahead.triage(
@@ -830,6 +834,18 @@ export let sort_streams_by_name = (matches: StreamPillData[], query: string): St
 
 export function rewire_sort_streams_by_name(value: typeof sort_streams_by_name): void {
     sort_streams_by_name = value;
+}
+
+export let sort_user_groups = (
+    matches: UserGroupPillData[],
+    query: string,
+): UserGroupPillData[] => {
+    const results = typeahead.triage(query, matches, (x) => x.name, compare_by_user_group_name);
+    return [...results.matches, ...results.rest];
+};
+
+export function rewire_sort_user_groups(value: typeof sort_user_groups): void {
+    sort_user_groups = value;
 }
 
 export function query_matches_person(