+++ zerver/tests/test_invite.py
@@ -28,7 +28,7 @@ from zerver.actions.create_realm import do_change_realm_subdomain, do_create_rea
 from zerver.actions.create_user import (
     do_create_user,
     process_new_human_user,
-    set_up_streams_for_new_human_user,
+    set_up_streams_and_groups_for_new_human_user,
 )
 from zerver.actions.default_streams import do_add_default_stream, do_remove_default_stream
 from zerver.actions.invites import (
@@ -43,7 +43,7 @@ from zerver.actions.realm_settings import (
     do_change_realm_plan_type,
     do_set_realm_property,
 )
-from zerver.actions.user_groups import check_add_user_group
+from zerver.actions.user_groups import check_add_user_group, do_change_user_group_permission_setting
 from zerver.actions.user_settings import do_change_full_name
 from zerver.actions.users import change_user_is_active
 from zerver.context_processors import common_context
@@ -56,6 +56,7 @@ from zerver.lib.send_email import FromAddress, deliver_scheduled_emails, send_fu
 from zerver.lib.streams import ensure_stream
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import find_key_by_email
+from zerver.lib.user_groups import get_direct_user_groups
 from zerver.models import (
     DefaultStream,
     Message,
@@ -112,7 +113,7 @@ class StreamSetupTest(ZulipTestCase):
         new_user = self.create_simple_new_user(realm, "alice@zulip.com")
 
         with self.assert_database_query_count(14):
-            set_up_streams_for_new_human_user(
+            set_up_streams_and_groups_for_new_human_user(
                 user_profile=new_user,
                 prereg_user=None,
                 default_stream_groups=[],
@@ -147,8 +148,40 @@ class StreamSetupTest(ZulipTestCase):
 
         new_user = self.create_simple_new_user(realm, new_user_email)
 
-        with self.assert_database_query_count(14):
-            set_up_streams_for_new_human_user(
+        with self.assert_database_query_count(15):
+            set_up_streams_and_groups_for_new_human_user(
+                user_profile=new_user,
+                prereg_user=prereg_user,
+                default_stream_groups=[],
+            )
+
+    def test_query_count_when_admin_assigns_groups(self) -> None:
+        admin = self.example_user("iago")
+        realm = admin.realm
+
+        hamletcharacters_group = NamedUserGroup.objects.get(name="hamletcharacters", realm=realm)
+        test_group = check_add_user_group(realm, "test", [admin], acting_user=admin)
+        user_groups = [hamletcharacters_group, test_group]
+
+        self.add_messages_to_stream("Rome")
+
+        new_user_email = "bob@zulip.com"
+
+        do_invite_users(
+            admin,
+            [new_user_email],
+            streams=[],
+            user_groups=user_groups,
+            include_realm_default_subscriptions=False,
+            invite_expires_in_minutes=1000,
+        )
+
+        prereg_user = PreregistrationUser.objects.get(email=new_user_email)
+
+        new_user = self.create_simple_new_user(realm, new_user_email)
+
+        with self.assert_database_query_count(10):
+            set_up_streams_and_groups_for_new_human_user(
                 user_profile=new_user,
                 prereg_user=prereg_user,
                 default_stream_groups=[],
@@ -179,6 +212,7 @@ class InviteUserBase(ZulipTestCase):
         stream_names: Sequence[str],
         notify_referrer_on_join: bool = True,
         invite_expires_in_minutes: int | None = INVITATION_LINK_VALIDITY_MINUTES,
+        group_ids: list[int] | None = None,
         body: str = "",
         invite_as: int = PreregistrationUser.INVITE_AS["MEMBER"],
         include_realm_default_subscriptions: bool = False,
@@ -191,6 +225,8 @@ class InviteUserBase(ZulipTestCase):
             newline separated.
 
         streams should be a list of strings.
+
+        group_ids should be a list of int.
         """
         stream_ids = [self.get_stream_id(stream_name, realm=realm) for stream_name in stream_names]
 
@@ -205,6 +241,7 @@ class InviteUserBase(ZulipTestCase):
                     "invitee_emails": invitee_emails,
                     "invite_expires_in_minutes": invite_expires_in,
                     "stream_ids": orjson.dumps(stream_ids).decode(),
+                    "group_ids": orjson.dumps(group_ids).decode() if group_ids else [],
                     "invite_as": invite_as,
                     "include_realm_default_subscriptions": orjson.dumps(
                         include_realm_default_subscriptions
@@ -742,6 +779,128 @@ class InviteUserTest(InviteUserBase):
             response, "Invalid invite_as: Value error, Not in the list of possible values"
         )
 
+    def test_invite_user_with_specified_user_groups_when_cannot_add_members(self) -> None:
+        hamlet = self.example_user("hamlet")
+        realm = hamlet.realm
+        # All users except guests have permission to send invites.
+        self.assertEqual(realm.can_invite_users_group.named_user_group.name, SystemGroups.MEMBERS)
+
+        nobody_group = NamedUserGroup.objects.get(
+            name=SystemGroups.NOBODY, realm=realm, is_system_group=True
+        )
+        test_group = check_add_user_group(
+            realm,
+            "test",
+            [hamlet],
+            acting_user=hamlet,
+            group_settings_map={
+                "can_manage_group": nobody_group,
+                "can_add_members_group": nobody_group,
+            },
+        )
+        hamletcharacters_group = NamedUserGroup.objects.get(name="hamletcharacters", realm=realm)
+
+        # Initialize settings with nobody allowed to add members or manage
+        # the group.
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            nobody_group,
+            acting_user=None,
+        )
+
+        do_change_user_group_permission_setting(
+            hamletcharacters_group,
+            "can_manage_group",
+            nobody_group,
+            acting_user=None,
+        )
+        do_change_user_group_permission_setting(
+            hamletcharacters_group,
+            "can_add_members_group",
+            nobody_group,
+            acting_user=None,
+        )
+
+        self.login("desdemona")
+        invitee = self.nonreg_email("test")
+        result = self.invite(invitee, [], group_ids=[test_group.id, hamletcharacters_group.id])
+        self.assert_json_error(result, "Insufficient permission")
+
+        # Test that user having permission to manage all groups can
+        # add user to groups through invitiation.
+        owners_group = NamedUserGroup.objects.get(
+            name=SystemGroups.OWNERS, realm=realm, is_system_group=True
+        )
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            owners_group,
+            acting_user=None,
+        )
+
+        self.login("iago")
+        result = self.invite(invitee, [], group_ids=[test_group.id, hamletcharacters_group.id])
+        self.assert_json_error(result, "Insufficient permission")
+
+        self.login("shiva")
+        result = self.invite(invitee, [], group_ids=[test_group.id, hamletcharacters_group.id])
+        self.assert_json_error(result, "Insufficient permission")
+
+        self.login("desdemona")
+
+        # Check that user does not have permission to add user to system groups
+        # even when having permission to manage all groups.
+        moderators_group = NamedUserGroup.objects.get(
+            name=SystemGroups.MODERATORS, realm=realm, is_system_group=True
+        )
+        result = self.invite(invitee, [], group_ids=[moderators_group.id])
+        self.assert_json_error(result, "Insufficient permission")
+
+        result = self.invite(invitee, [], group_ids=[test_group.id, hamletcharacters_group.id])
+        self.assert_json_success(result)
+        self.assertTrue(find_key_by_email(invitee))
+
+        # Test that user having permission to add members to a group can
+        # add user to that group through invitiation.
+        do_change_user_group_permission_setting(
+            test_group,
+            "can_add_members_group",
+            moderators_group,
+            acting_user=None,
+        )
+        self.login("hamlet")
+        invitee = self.nonreg_email("bob")
+        result = self.invite(invitee, [], group_ids=[test_group.id, hamletcharacters_group.id])
+        self.assert_json_error(result, "Insufficient permission")
+
+        self.login("shiva")
+        result = self.invite(invitee, [], group_ids=[test_group.id, hamletcharacters_group.id])
+        self.assert_json_error(result, "Insufficient permission")
+
+        result = self.invite(invitee, [], group_ids=[test_group.id])
+        self.assert_json_success(result)
+        self.assertTrue(find_key_by_email(invitee))
+
+        # Test that user having permission to manage a group can
+        # add user to that group through invitiation.
+        do_change_user_group_permission_setting(
+            hamletcharacters_group,
+            "can_manage_group",
+            moderators_group,
+            acting_user=None,
+        )
+        invitee = self.nonreg_email("alice")
+
+        self.login("hamlet")
+        result = self.invite(invitee, [], group_ids=[test_group.id, hamletcharacters_group.id])
+        self.assert_json_error(result, "Insufficient permission")
+
+        self.login("shiva")
+        result = self.invite(invitee, [], group_ids=[hamletcharacters_group.id, test_group.id])
+        self.assert_json_success(result)
+        self.assertTrue(find_key_by_email(invitee))
+
     def test_successful_invite_user_as_guest_from_normal_account(self) -> None:
         self.login("hamlet")
         invitee = self.nonreg_email("alice")
@@ -850,6 +1009,29 @@ class InviteUserTest(InviteUserBase):
         # passed in streams list.
         self.submit_reg_form_for_user(invitee, "password")
         self.check_user_subscribed_only_to_streams("test1", [denmark, sandbox, verona, zulip])
+
+    def test_successful_invite_users_with_specified_user_groups(self) -> None:
+        invitee = self.nonreg_email("bob")
+        iago = self.example_user("iago")
+        self.login("iago")
+
+        user_group1 = check_add_user_group(iago.realm, "test1", [], acting_user=iago)
+        user_group2 = check_add_user_group(iago.realm, "test2", [], acting_user=iago)
+
+        group_ids = [user_group1.id, user_group2.id]
+
+        self.assert_json_success(self.invite(invitee, [], group_ids=group_ids))
+        self.assertTrue(find_key_by_email(invitee))
+        self.submit_reg_form_for_user(invitee, "password")
+
+        # bob is a direct member of two role-based system groups also.
+        user_groups_subscriptions = get_direct_user_groups(self.nonreg_user("bob"))
+        user_group_names = [group.named_user_group.name for group in user_groups_subscriptions]
+
+        self.assertEqual(
+            set(user_group_names),
+            {"test1", "test2", SystemGroups.MEMBERS, SystemGroups.FULL_MEMBERS},
+        )
 
     def test_invite_others_to_realm_setting(self) -> None:
         """
@@ -1219,6 +1401,17 @@ earl-test@zulip.com""",
         self.assert_json_error(
             self.invite("iago-test@zulip.com", ["NotARealStream"]),
             f"Invalid channel ID {self.INVALID_STREAM_ID}. No invites were sent.",
+        )
+        self.check_sent_emails([])
+
+    def test_invalid_user_group(self) -> None:
+        """
+        Tests inviting to a non-existent user group.
+        """
+        self.login("hamlet")
+        self.assert_json_error(
+            self.invite("iago-test@zulip.com", ["Denmark"], group_ids=[5678]),
+            "Invalid user group",
         )
         self.check_sent_emails([])
 
@@ -2528,6 +2721,7 @@ class MultiuseInviteTest(ZulipTestCase):
         self,
         streams: list[Stream] | None = None,
         date_sent: datetime | None = None,
+        user_groups: list[NamedUserGroup] | None = None,
         include_realm_default_subscriptions: bool = False,
     ) -> str:
         invite = MultiuseInvite(
@@ -2539,6 +2733,9 @@ class MultiuseInviteTest(ZulipTestCase):
 
         if streams is not None:
             invite.streams.set(streams)
+
+        if user_groups is not None:
+            invite.groups.set(user_groups)
 
         if date_sent is None:
             date_sent = timezone_now()
@@ -2682,6 +2879,27 @@ class MultiuseInviteTest(ZulipTestCase):
             name5, [rome, default_streams[0], default_streams[1], default_streams[2]]
         )
 
+    def test_multiuse_link_with_specified_user_groups(self) -> None:
+        iago = self.example_user("iago")
+        self.login("iago")
+
+        user_group1 = check_add_user_group(iago.realm, "test1", [], acting_user=iago)
+        user_group2 = check_add_user_group(iago.realm, "test2", [], acting_user=iago)
+
+        user_groups = [user_group1, user_group2]
+
+        invite_link = self.generate_multiuse_invite_link(user_groups=user_groups)
+        self.check_user_able_to_register(self.nonreg_email("bob"), invite_link)
+
+        # bob is a direct member of two role-based system groups also.
+        user_groups_subscriptions = get_direct_user_groups(self.nonreg_user("bob"))
+        user_group_names = [group.named_user_group.name for group in user_groups_subscriptions]
+
+        self.assertEqual(
+            set(user_group_names),
+            {"test1", "test2", SystemGroups.MEMBERS, SystemGroups.FULL_MEMBERS},
+        )
+
     def test_multiuse_link_different_realms(self) -> None:
         """
         Verify that an invitation generated for one realm can't be used
@@ -2795,6 +3013,54 @@ class MultiuseInviteTest(ZulipTestCase):
         self.assert_length(get_default_streams_for_realm_as_dicts(self.realm.id), 3)
         self.check_user_subscribed_only_to_streams("alice", [])
 
+    def test_create_multiuse_link_with_specified_user_groups_api_call(self) -> None:
+        iago = self.example_user("iago")
+        self.login("iago")
+
+        user_group1 = check_add_user_group(iago.realm, "test1", [], acting_user=iago)
+        user_group2 = check_add_user_group(iago.realm, "test2", [], acting_user=iago)
+
+        group_ids = [user_group1.id, user_group2.id]
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "group_ids": orjson.dumps(group_ids).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+            },
+        )
+        invite_link = self.assert_json_success(result)["invite_link"]
+        self.check_user_able_to_register(self.nonreg_email("bob"), invite_link)
+
+        # bob is a direct member of two role-based system groups also.
+        user_groups_subscriptions = get_direct_user_groups(self.nonreg_user("bob"))
+        user_group_names = [group.named_user_group.name for group in user_groups_subscriptions]
+
+        self.assertEqual(
+            set(user_group_names),
+            {"test1", "test2", SystemGroups.MEMBERS, SystemGroups.FULL_MEMBERS},
+        )
+
+        self.login("iago")
+        group_ids = []
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "group_ids": orjson.dumps(group_ids).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+            },
+        )
+        invite_link = self.assert_json_success(result)["invite_link"]
+        self.check_user_able_to_register(self.nonreg_email("newuser"), invite_link)
+
+        # bob is a direct member of two role-based system groups also.
+        user_groups_subscriptions = get_direct_user_groups(self.nonreg_user("newuser"))
+        user_group_names = [group.named_user_group.name for group in user_groups_subscriptions]
+
+        self.assertEqual(
+            set(user_group_names),
+            {SystemGroups.MEMBERS, SystemGroups.FULL_MEMBERS},
+        )
+
     def test_multiuse_invite_without_permission_to_subscribe_others(self) -> None:
         realm = get_realm("zulip")
         members_group = NamedUserGroup.objects.get(
@@ -2858,6 +3124,131 @@ class MultiuseInviteTest(ZulipTestCase):
             },
         )
         self.assert_json_success(result)
+
+    def test_multiuser_link_with_specified_user_groups_when_cannot_add_members(self) -> None:
+        hamlet = self.example_user("hamlet")
+        realm = hamlet.realm
+        # All users except guests have permission to create multiuse invite.
+        members_group = NamedUserGroup.objects.get(
+            name=SystemGroups.MEMBERS, realm=realm, is_system_group=True
+        )
+        do_change_realm_permission_group_setting(
+            realm, "create_multiuse_invite_group", members_group, acting_user=None
+        )
+
+        nobody_group = NamedUserGroup.objects.get(
+            name=SystemGroups.NOBODY, realm=realm, is_system_group=True
+        )
+        test_group = check_add_user_group(
+            realm,
+            "test",
+            [hamlet],
+            acting_user=hamlet,
+            group_settings_map={
+                "can_manage_group": nobody_group,
+                "can_add_members_group": nobody_group,
+            },
+        )
+        hamletcharacters_group = NamedUserGroup.objects.get(name="hamletcharacters", realm=realm)
+
+        def check_create_multiuse_invite(
+            user: str, group_ids: list[int], error_msg: str | None = None
+        ) -> None:
+            self.login(user)
+            result = self.client_post(
+                "/json/invites/multiuse",
+                {
+                    "group_ids": orjson.dumps(group_ids).decode(),
+                    "invite_expires_in_minutes": 2 * 24 * 60,
+                },
+            )
+            if error_msg is not None:
+                self.assert_json_error(result, error_msg)
+            else:
+                self.assert_json_success(result)
+
+        # Initialize settings with nobody allowed to add members or manage
+        # the group.
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            nobody_group,
+            acting_user=None,
+        )
+        do_change_user_group_permission_setting(
+            hamletcharacters_group,
+            "can_manage_group",
+            nobody_group,
+            acting_user=None,
+        )
+        do_change_user_group_permission_setting(
+            hamletcharacters_group,
+            "can_add_members_group",
+            nobody_group,
+            acting_user=None,
+        )
+
+        check_create_multiuse_invite(
+            "desdemona", [test_group.id, hamletcharacters_group.id], "Insufficient permission"
+        )
+
+        # Test that user having permission to manage all groups can
+        # add users to groups through invitiation.
+        owners_group = NamedUserGroup.objects.get(
+            name=SystemGroups.OWNERS, realm=realm, is_system_group=True
+        )
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            owners_group,
+            acting_user=None,
+        )
+
+        check_create_multiuse_invite(
+            "iago", [test_group.id, hamletcharacters_group.id], "Insufficient permission"
+        )
+        check_create_multiuse_invite(
+            "shiva", [test_group.id, hamletcharacters_group.id], "Insufficient permission"
+        )
+
+        # Check that user does not have permission to add user to system groups
+        # even when having permission to manage all groups.
+        moderators_group = NamedUserGroup.objects.get(
+            name=SystemGroups.MODERATORS, realm=realm, is_system_group=True
+        )
+        check_create_multiuse_invite("desdemona", [moderators_group.id], "Insufficient permission")
+        check_create_multiuse_invite("desdemona", [test_group.id, hamletcharacters_group.id])
+
+        # Test that user having permission to add members to a group can
+        # add user to that group through invitiation.
+        do_change_user_group_permission_setting(
+            test_group,
+            "can_add_members_group",
+            moderators_group,
+            acting_user=None,
+        )
+        check_create_multiuse_invite(
+            "hamlet", [test_group.id, hamletcharacters_group.id], "Insufficient permission"
+        )
+
+        check_create_multiuse_invite(
+            "shiva", [test_group.id, hamletcharacters_group.id], "Insufficient permission"
+        )
+        check_create_multiuse_invite("shiva", [test_group.id])
+
+        # Test that user having permission to manage a group can
+        # add user to that group through invitiation.
+        do_change_user_group_permission_setting(
+            hamletcharacters_group,
+            "can_manage_group",
+            moderators_group,
+            acting_user=None,
+        )
+
+        check_create_multiuse_invite(
+            "hamlet", [test_group.id, hamletcharacters_group.id], "Insufficient permission"
+        )
+        check_create_multiuse_invite("shiva", [test_group.id, hamletcharacters_group.id])
 
     def test_create_multiuse_invite_group_setting(self) -> None:
         realm = get_realm("zulip")
@@ -3047,6 +3438,17 @@ class MultiuseInviteTest(ZulipTestCase):
             },
         )
         self.assert_json_error(result, "Invalid channel ID 54321. No invites were sent.")
+
+    def test_create_multiuse_link_invalid_user_group_api_call(self) -> None:
+        self.login("iago")
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "group_ids": orjson.dumps([5438]).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+            },
+        )
+        self.assert_json_error(result, "Invalid user group")
 
     def test_create_multiuse_link_invalid_invite_as_api_call(self) -> None:
         self.login("iago")